refactor: ratelimit パッケージの責務分離・命名・配置を整理
- httpratelimit/ratelimit.go を limiter.go にリネーム（パッケージ名と同名ファイル解消）
- clientratelimit を internal/platform/ から internal/shared/ に移動し rate_limiter.go を limiter.go にリネーム（usecase からの利用を depguard と整合）
- clientratelimit.RateLimiterInterface を削除し、candles/usecase と symbollist/usecase に RateLimiter インターフェースを定義（インターフェースは利用者側で定義する Go 慣例に追従）
- httpratelimit.ScriptHash を非公開化し、外部テスト向けに ExpectAllow テストヘルパーを mockhelpers.go に追加
- CLAUDE.md / README.md / candles/README.md のディレクトリ構造記述と Mermaid 図を更新
