+++ cmd/logo-ingest/main.go
@@ -10,8 +10,8 @@ import (
 	"stock_backend/internal/app/di"
 	symbollistadapters "stock_backend/internal/feature/symbollist/adapters"
 	symbollistusecase "stock_backend/internal/feature/symbollist/usecase"
-	"stock_backend/internal/platform/clientratelimit"
 	"stock_backend/internal/platform/db"
+	"stock_backend/internal/shared/clientratelimit"
 )
 
 const (
